deo sua nua

+++ Sources/engine/DataType.h
@@ -41,7 +41,7 @@ typedef signed char Int8;
 typedef unsigned char Uint8;
 typedef unsigned short Uint16;
 typedef unsigned int Uint32;
-typedef unsigned long int Uint64;
+typedef unsigned long long int Uint64;
 
 enum HoolCodeState {
     vDoNothing = 0, //do not do anything
